feat: introduce AnyCodable for flexible JSON decoding
