+++ Sources/WearServiceAPI.swift
@@ -591,80 +591,3 @@ internal struct GarminWebhookUrlsResponse: Codable {
     let endpoints: [String: String]
 }
 
-// MARK: - Supporting Types
-
-/// Type-erased Codable for flexible JSON decoding
-internal struct AnyCodable: Codable {
-    let value: Any
-    
-    init(_ value: Any) {
-        self.value = value
-    }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        
-        // Handle null values first
-        if container.decodeNil() {
-            // Use NSNull to represent null values (which is what JSON null maps to)
-            value = NSNull()
-            return
-        }
-        
-        // Try decoding different types in order
-        if let bool = try? container.decode(Bool.self) {
-            value = bool
-        } else if let int = try? container.decode(Int.self) {
-            value = int
-        } else if let double = try? container.decode(Double.self) {
-            value = double
-        } else if let string = try? container.decode(String.self) {
-            value = string
-        } else if let array = try? container.decode([AnyCodable].self) {
-            value = array.map { $0.value }
-        } else if let dict = try? container.decode([String: AnyCodable].self) {
-            value = dict.mapValues { $0.value }
-        } else {
-            throw DecodingError.dataCorruptedError(
-                in: container,
-                debugDescription: "AnyCodable value cannot be decoded"
-            )
-        }
-    }
-    
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-        
-        // Handle null values
-        if value is NSNull {
-            try container.encodeNil()
-            return
-        }
-        
-        switch value {
-        case let bool as Bool:
-            try container.encode(bool)
-        case let int as Int:
-            try container.encode(int)
-        case let double as Double:
-            try container.encode(double)
-        case let string as String:
-            try container.encode(string)
-        case let array as [Any]:
-            let codableArray = array.map { AnyCodable($0) }
-            try container.encode(codableArray)
-        case let dict as [String: Any]:
-            let codableDict = dict.mapValues { AnyCodable($0) }
-            try container.encode(codableDict)
-        default:
-            throw EncodingError.invalidValue(
-                value,
-                EncodingError.Context(
-                    codingPath: container.codingPath,
-                    debugDescription: "AnyCodable value cannot be encoded"
-                )
-            )
-        }
-    }
-}
-
